removed bug in conf.py
