+++ docs/source/conf.py
@@ -85,21 +85,22 @@ sphinx_gallery_conf = {
     'matplotlib_animations': True,
     'image_srcset': ["2x"],
 
-    'within_subsection_order': ExplicitOrder(['../../scripts/plot',
-                                       '../../scripts/scatter',
-                                       '../../scripts/imshow',
-                                       '../../scripts/bar_chart',
-                                       '../../scripts/hist',
-                                       '../../scripts/lollipop_plot',
-                                       '../../scripts/dumbell_plot',
-                                       '../../scripts/regplot',
-                                       '../../scripts/ridge',
-                                       '../../scripts/pie',
-                                       '../../scripts/contour',
-                                       '../../scripts/circular_bar_plot',
-                                       '../../scripts/parallel_coordinates',
-                                       '../../scripts/spider_plot',
-                                       '../../scripts/taylor_plot',
+    'within_subsection_order': ExplicitOrder([
+        '../../examples/plot',
+        '../../examples/scatter',
+        '../../examples/imshow',
+        '../../examples/bar_chart',
+        '../../examples/hist',
+        '../../examples/lollipop_plot',
+        '../../examples/dumbell_plot',
+        '../../examples/regplot',
+        '../../examples/ridge',
+        '../../examples/pie',
+        '../../examples/contour',
+        '../../examples/circular_bar_plot',
+        '../../examples/parallel_coordinates',
+        '../../examples/spider_plot',
+        '../../examples/taylor_plot',
                                        ]),
 }
