+++ easy_mpl/_main.py
@@ -11,14 +11,11 @@ __all__ = [
     "dumbbell_plot",
     "ridge",
     "parallel_coordinates",
-    "lollipop_plot",
-    "circular_bar_plot",
-    "spider_plot",
+    "lollipop_plot"
 ]
 
-import math
 import random
-from typing import Union, Tuple, List, Optional
+from typing import Union, Tuple, List
 
 import numpy as np
 import pandas as pd
@@ -32,7 +29,7 @@ import matplotlib.gridspec as grid_spec
 from mpl_toolkits.axes_grid1 import make_axes_locatable
 
 from .utils import kde
-from .utils import register_projections
+from .utils import _rescale
 from .utils import to_1d_array, make_cols_from_cmap, _regplot, process_axis
 from .utils import BAR_CMAPS, regplot_combs, RIDGE_CMAPS, annotate_imshow
 
@@ -1436,12 +1433,6 @@ def is_categorical(array) -> bool:
     return not np.issubdtype(array.dtype, np.number)
 
 
-def _rescale(y, _min=0.0, _max=1.0):
-    y_std = (y - np.min(y, axis=0)) / (np.max(y, axis=0) - np.min(y, axis=0))
-
-    return y_std * (_max - _min) + _min
-
-
 def label_encoder(arr):
     # label encoder of numpy array with categorical values
     return np.unique(arr, return_inverse=True)[1]
@@ -1581,347 +1572,4 @@ def _lollipop_horizontal(ax, x, y, line_style, line_color, line_width, line_kws,
                s=marker_size, **marker_kws)
     ax.hlines(x, np.zeros(len(y)), y, color=line_color,
               linestyle=line_style, linewidth=line_width, **line_kws)
-    return ax
-
-
-def circular_bar_plot(
-        data,
-        labels: list = None,
-        sort=False,
-        color: Union[str, list, np.ndarray] = None,
-        label_format: str = None,
-        min_max_range: tuple = None,
-        label_padding: int = 4,
-        figsize: tuple = None,
-        show: bool = True,
-        text_kws: dict = None,
-        **kwargs
-) -> plt.Axes:
-    """
-    Plot a circular bar plot.
-
-    Parameters
-    ----------
-    data : list, np.ndarray, pd.Series, dict
-        Data to plot. If it is a dictionary, then its keys will be used
-        as labels and values will be used as data.
-    labels : list, optional
-        Labels for each data point.
-    sort : bool, optional
-        Sort the data by the values.
-    color : str, list, np.ndarray, optional
-        Color for each data point. It can be a single color or a colormap from
-        plt.colormaps.
-    label_format : str, optional
-        Format for the labels.
-    min_max_range : tuple, optional
-        Minimum and maximum range for normalizing the data.
-    label_padding : int, optional
-        space between the labels and the bars.
-    figsize : tuple, optional
-        Size of the figure.
-    show : bool, optional
-        Show the plot.
-    **kwargs : optional
-        Additional keyword arguments to pass to the process_axis function.
-
-    Returns
-    -------
-    ax : plt.Axes
-        Axes of the plot.
-
-    Note
-    ----
-        If nan values are present in the data, they will be ignored.
-
-    Examples
-    --------
-    >>> import numpy as np
-    >>> from easy_mpl import circular_bar_plot
-    >>> data = np.random.random(50, )
-    ... # basic
-    >>> circular_bar_plot(data)
-    ... # with names
-    >>> names = [f"{i}" for i in range(50)]
-    >>> circular_bar_plot(data, names)
-    ... # sort values
-    >>> circular_bar_plot(data, names, sort=True)
-    ... # custom color map
-    >>> circular_bar_plot(data, names, color='viridis')
-    ... # custom min and max range
-    >>> circular_bar_plot(data, names, min_max_range=(1, 10), label_padding=1)
-    ... # custom label format
-    >>> circular_bar_plot(data, names, label_format='{} {:.4f}')
-
-    """
-
-    text_kws = text_kws or {}
-
-    plt.close('all')
-    plt.figure(figsize=figsize or (8, 12))
-    ax = plt.subplot(111, polar=True)
-    plt.axis('off')
-
-    if isinstance(data, pd.DataFrame):
-        values = data.values
-    elif isinstance(data, dict):
-        values = np.array(list(data.values()))
-        labels = labels or list(data.keys())
-    else:
-        data = to_1d_array(data)
-        values = data
-
-    if labels is None:
-        labels = ['' for _ in range(len(values))]
-        label_format = label_format or "{} {:.2f}"
-    else:
-        label_format = label_format or "{}: {:.2f}"
-
-    # remove nan values
-    val_nan_idx = np.isnan(values)
-    if val_nan_idx.any():
-        values = values[~val_nan_idx]
-        labels = [labels[i] for i in range(len(labels)) if not val_nan_idx[i]]
-
-    if color is None:
-        color = make_cols_from_cmap(random.choice(BAR_CMAPS), len(values), 0.2)
-    elif isinstance(color, str) and color in plt.colormaps():
-        color = make_cols_from_cmap(color, len(values), 0.2)
-    else:
-        color = color
-
-    assert len(values) == len(labels)
-
-    min_max_range = min_max_range or (30, 100)
-    lower_limit = min_max_range[0]
-    heights = _rescale(values.reshape(-1, 1), lower_limit, min_max_range[1]).reshape(-1, )
-
-    if sort:
-        sort_idx = np.argsort(heights)
-        heights = heights[sort_idx]
-        labels = [labels[i] for i in sort_idx]
-        values = values[sort_idx]
-        # color = color[sort_idx]
-
-    # Compute the width of each bar. In total we have 2*Pi = 360°
-    width = 2 * np.pi / len(heights)
-
-    # Compute the angle each bar is centered on:
-    indexes = list(range(1, len(heights) + 1))
-    angles = [element * width for element in indexes]
-
-    # Draw bars
-    bars = ax.bar(
-        x=angles,
-        height=heights,
-        width=width,
-        bottom=lower_limit,
-        linewidth=2,
-        edgecolor="white",
-        color=color,
-    )
-
-    # Add labels
-    for bar, angle, label, val in zip(bars, angles, labels, values):
-
-        label = label_format.format(label, val)
-
-        # Labels are rotated. Rotation must be specified in degrees :(
-        rotation = np.rad2deg(angle)
-
-        # Flip some labels upside down
-        if angle >= np.pi / 2 and angle < 3 * np.pi / 2:
-            alignment = "right"
-            rotation = rotation + 180
-        else:
-            alignment = "left"
-
-        # Finally add the labels
-        ax.text(
-            x=angle,
-            y=lower_limit + bar.get_height() + label_padding,
-            s=label,
-            ha=alignment,
-            va='center',
-            rotation=rotation,
-            rotation_mode="anchor",
-            **text_kws
-        )
-
-    if kwargs:
-        process_axis(ax, **kwargs)
-
-    if show:
-        plt.show()
-
-    return ax
-
-
-def spider_plot(
-        values:Union[pd.DataFrame, np.ndarray, list],
-        tick_labels:list = None,
-        highlight:Union[int, float]=None,
-        plot_kws: Optional[Union[dict, List[dict]]] = None,
-        xtick_kws: Optional[dict] = None,
-        fill_kws : Optional[dict] = None,
-        frame:str = "circle",
-        color : Union[List[str], str] = None,
-        fill_color: Union[List[str], str] = None,
-        leg_kws : dict = None,
-        labels: list = None,
-        show: Optional[bool] = True,
-        figsize=None,
-)->plt.Axes:
-    """
-    Draws spider plot on an axes
-
-    Parameters
-    ----------
-        values :
-            values to display.
-        tick_labels : list, optional
-            tick labels. It's length should be equal to length of values
-        plot_kws : dict, optional
-            These can include, ``color``, ``linewidth``, ``linestyle`` etc
-        xtick_kws : dict, optional
-            These can include ``color``, ``size`` etc.
-        fill_kws : dict, optional
-            These can include ``color``, ``alpha`` etc.
-        highlight : int/float optional (default=None)
-            whether to highlight a certain circular line or not
-        color : str
-            colormap to use
-        frame: str, optional (default="circle")
-            whether the outer frame and grids should be polygon or circle
-        ax : plt.Axes
-            matplotlib axes on which to draw the axes
-        figsize : tuple, optional (default=None)
-            figure size
-        fill_color :
-            color to use for filling
-        leg_kws : dict
-            keyword arguments that will go to ax.legend()
-        labels: list, optional (default=None)
-            the labels for values
-        show : bool, optional (default=True)
-            whether to show the plot or not
-
-    Returns
-    -------
-    plt.Axes
-        matplotlib axes on which plot is drawn
-
-    Examples
-    --------
-    >>> from easy_mpl import spider_plot
-    >>> values = [-0.2, 0.1, 0.0, 0.1, 0.2, 0.3]
-    >>> spider_plot(values=values)
-    ... # specifying labels
-    labels = ['a', 'b','c', 'd', 'e', 'f']
-    >>> spider_plot(values=values, labels=labels)
-    ... # specifying tick size
-    >>> spider_plot(values, labels, xtick_kws={'size': 13})
-    ...
-    >>> df = pd.DataFrame.from_dict(
-    ... {'summer': {'a': -0.2, 'b': 0.1, 'c': 0.0, 'd': 0.1, 'e': 0.2, 'f': 0.3},
-    ... 'winter': {'a': -0.3, 'b': 0.1, 'c': 0.0, 'd': 0.2, 'e': 0.15,'f': 0.25}})
-    >>> spider_plot(df, xtick_kws={'size': 13})
-    ... # use polygon frame
-    >>> spider_plot(values=values, frame="polygon")
-
-    """
-
-    # todo, allow grids and frame to have different types
-
-    if isinstance(values, (list, tuple)):
-        values = np.array(values).reshape(-1, 1)
-
-    if tick_labels is None:
-        if isinstance(values, (pd.DataFrame, pd.Series)):
-            tick_labels = values.index
-        else:
-            tick_labels = [f"F{i}" for i in range(len(values))]
-
-    if labels is None:
-        if isinstance(values, pd.Series):
-            labels = [values.name]
-        elif isinstance(values, pd.DataFrame):
-            labels = values.columns.tolist()
-        else:
-            labels = [f'Value_{i}' for i in range(values.shape[1])]
-
-    if isinstance(values, (pd.DataFrame, pd.Series)):
-        values = values.values
-
-    N = len(tick_labels)
-    assert N == len(values)
-
-    if not isinstance(plot_kws, list):
-        plot_kws = [plot_kws for _ in range(N)]
-
-    def_color = plt.cm.get_cmap('Set2', values.shape[1])
-
-    if color is None:
-        color = [def_color(i) for i in range(values.shape[1])]
-    if fill_color is None:
-        fill_color = [def_color(i) for i in range(values.shape[1])]
-
-    if not isinstance(color, list):
-        color = [color for _ in range(N)]
-
-    if not isinstance(fill_color, list):
-        fill_color = [fill_color for _ in range(N)]
-
-    angles = [n / float(N) * 2 * math.pi for n in range(N)]
-    angles += angles[:1]
-
-    if frame == "polygon":
-        register_projections(N, frame)
-        fig, ax = plt.subplots(figsize=figsize, nrows=1, ncols=1,
-                                subplot_kw= dict(projection='radar'))
-    else:
-        ax = plt.subplot(polar=True)
-
-    _xtick_kws = {'color': 'grey', 'size': 8}
-    xtick_kws = xtick_kws or {}
-    _xtick_kws.update(xtick_kws)
-
-    plt.xticks(angles[:-1], tick_labels, **_xtick_kws)
-
-    for idx in range(values.shape[1]):
-
-        plot_kws_ = plot_kws[idx]
-
-        _plot_kws = {"color": color[idx], "linewidth": 2, "linestyle": 'solid'}
-        plot_kws_ = plot_kws_ or {}
-        _plot_kws.update(plot_kws_)
-
-        val = values[:, idx].tolist()
-        val.append(val[0])
-        ax.plot(angles, val, **_plot_kws)
-
-        _fill_kws = {"color":fill_color[idx], "alpha":.4, 'label': '_nolegend_'}
-        fill_kws = fill_kws or {}
-        _fill_kws.update(fill_kws)
-        ax.fill(angles, val, **_fill_kws)
-
-    if highlight:
-        val_minus = highlight - 0.005
-        val_plus = highlight + 0.005
-        ax.fill_between(np.linspace(0, 2 * np.pi, 100),
-                        val_minus,
-                        val_plus,
-                        color='red',
-                        zorder=10)
-
-    plt.gca().set_rmax(.4)
-
-    _leg_kws = {'loc': 'best', 'labelspacing': 0.1, 'fontsize': 'small'}
-    leg_kws = leg_kws or {}
-    _leg_kws.update(leg_kws)
-    legend = ax.legend(labels, **_leg_kws)
-
-    if show:
-        plt.show()
-
     return ax
