+++ tests/test_main.py
@@ -18,9 +18,6 @@ from easy_mpl.utils import BAR_CMAPS, make_cols_from_cmap
 from easy_mpl import dumbbell_plot, ridge
 from easy_mpl import parallel_coordinates
 from easy_mpl import lollipop_plot
-from easy_mpl import circular_bar_plot
-from easy_mpl import spider_plot
-
 
 
 def get_chart_data(n):
@@ -828,131 +825,6 @@ class TestLollipopPlot(unittest.TestCase):
         y = np.random.randint(0, 20, size=10).astype("float32")
         y[3] = np.nan
         ax = lollipop_plot(y, title="with nan vals", show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-
-class TestCircularBarPlot(unittest.TestCase):
-    show = False
-    data = np.random.random(50, )
-    names = [f"{i}" for i in range(50)]
-
-    def test_basic(self):
-        ax = circular_bar_plot(self.data, title="basic", show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_custom_color(self):
-        ax = circular_bar_plot(self.data, color="#61a4b2", show=self.show,
-                         title="custom color")
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_with_names(self):
-        ax = circular_bar_plot(self.data, self.names, title="with names",
-                               show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_custom_fs(self):
-        ax = circular_bar_plot(self.data, self.names, title="custon_fontsize",
-                               text_kws={"fontsize": 16}, show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_sort(self):
-        ax = circular_bar_plot(self.data, self.names, sort=True,
-                               title="sort", show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_cmap(self):
-        ax = circular_bar_plot(self.data, self.names, color='viridis',
-                               show=self.show, title="cmap")
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_minmax_range(self):
-
-        ax = circular_bar_plot(self.data, self.names, min_max_range=(1, 10),
-                              label_padding=1, show=self.show, title="minmax range")
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_custom_label_format(self):
-        ax = circular_bar_plot(self.data, self.names, label_format='{} {:.4f}',
-                               show=self.show, title="custom label format")
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_multiple_values(self):
-        # circular_bar_plot(data, names)
-        # assert isinstance(ax, plt.Axes)
-        return
-
-    def test_with_nan_vals(self):
-        data = np.random.random(50)
-        data[10] = np.nan
-        names = [f"{i}" for i in range(50)]
-        ax = circular_bar_plot(data, names, show=self.show, title="with nan vals")
-        assert isinstance(ax, plt.Axes)
-        return
-
-
-class TestSpiderPlot(unittest.TestCase):
-    values = [-0.2, 0.1, 0.0, 0.1, 0.2, 0.3]
-    labels = ['a', 'b', 'c', 'd', 'e', 'f']
-
-    df = pd.DataFrame.from_dict(
-        {'a': {'a': -0.2, 'b': 0.1, 'c': 0.0, 'd': 0.1, 'e': 0.2, 'f': 0.3},
-         'b': {'a': -0.3, 'b': 0.1, 'c': 0.0, 'd': 0.2, 'e': 0.15, 'f': 0.25},
-         'c': {'a': -0.1, 'b': 0.3, 'c': 0.15, 'd': 0.24, 'e': 0.18, 'f': 0.2}})
-
-    show = False
-
-    def test_basic(self):
-        ax = spider_plot(values=self.values, show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_specify_tick_labels(self):
-
-
-        ax = spider_plot(values=self.values,
-                         labels=self.labels, show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_tick_size(self):
-
-        ax = spider_plot(self.values, self.labels, xtick_kws={'size': 13},
-                         show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_polygon_frame_single(self):
-
-        ax = spider_plot(values=self.values, frame="polygon", show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_dict(self):
-
-        ax = spider_plot(self.df, xtick_kws={'size': 13}, show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_dict_polygon(self):
-
-        ax = spider_plot(self.df, xtick_kws={'size': 13}, frame="polygon",
-                    color=['b', 'r', 'g', 'm'],
-                    fill_color=['b', 'r', 'g', 'm'],
-                         show=self.show)
-        assert isinstance(ax, plt.Axes)
-        return
-
-    def test_2d_np_array(self):
-        ax = spider_plot(self.df.values, labels=self.df.columns.tolist(), show=self.show)
         assert isinstance(ax, plt.Axes)
         return
